Removing SitePickerSearchView because SitePickerActivity dependency wasn't desired
